requirements.txt
Django==1.11
django-filter==1.1.0
django-hosts==3.0
djangorestframework==3.7.0
Markdown==2.6.11
Pillow==5.1.0
pkg-resources==0.0.0
pytz==2018.4
reportlab==3.4.0
